src/requirements.txt
requests
gitpython
pyyaml